+++ common/include/ag_version.h
@@ -1,2 +1,2 @@
 #pragma once
-#define AG_DNSLIBS_VERSION "1.5.46"
+#define AG_DNSLIBS_VERSION "1.6.0"